fix typo

+++ Ro.Inventario.Web/Repos/RolesRepo.cs
@@ -89,6 +89,6 @@ public class RolesRepo : IRolesRepo
 		);
 
 		var count = await Db.ExecuteScalar(cmd);
-		return ((int)count) > 0;
+		return ((long)count) > 0;
 	}
 }
